objectDetection/requirements.txt
numpy==1.18.5
oprncv-python==4.5.1.48
onnx==1.9.0
onnxruntime-gpu==1.8.1
scipy==1.4.1
pillow==8.1.2